+++ src/app/app.module.ts
@@ -1,21 +1,23 @@
 import { NgModule, LOCALE_ID } from '@angular/core';
-
 import { BrowserModule } from '@angular/platform-browser';
 import { AngularFireModule } from '@angular/fire';
 import { AngularFirestoreModule } from '@angular/fire/firestore';
 import { ServiceWorkerModule } from '@angular/service-worker';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FlexLayoutModule } from '@angular/flex-layout';
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
+import { registerLocaleData } from '@angular/common';
+import localeEs from '@angular/common/locales/es';
 
 import { environment } from '@environments/environment';
 
+import { SpinnerInterceptor } from '@shared/interceptors/spinner.interceptor';
+import { SharedModule } from '@app/shared/shared.module';
+import { HomeModule } from '@app/home/home.module';
+
 import { AppRoutingModule } from '@app/app-routing.module';
 import { AppComponent } from '@app/app.component';
-import { SharedModule } from '@app/shared/shared.module';
 
-import { registerLocaleData } from '@angular/common';
-import localeEs from '@angular/common/locales/es';
-import { HomeModule } from './home/home.module';
 registerLocaleData(localeEs, 'es');
 
 @NgModule({
@@ -34,6 +36,7 @@ registerLocaleData(localeEs, 'es');
     HomeModule
   ],
   providers: [
+    { provide: HTTP_INTERCEPTORS, useClass: SpinnerInterceptor, multi: true },
     { provide: LOCALE_ID, useValue: 'es' },
   ],
   bootstrap: [AppComponent]