+++ changelog.md
@@ -12,10 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Add
 
 - Alerted notice will appeared as a modal at home component.
+- Add links module, for using instead of news module.
 
 ### Change
 
 - Refactor stylesheets and add one for dialogs.
+- Add loading component with interceptor.
 
 ## [0.0.10] - 2021-03-11 Styling - Presentación al Ayuntamiento